fix up touch to be async!
