+++ crates/utiles/src/cli/commands/dev.rs
@@ -1,3 +1,4 @@
+#![allow(clippy::unwrap_used)]
 use clap::Parser;
 use tracing::{debug, warn};
