+++ crates/utiles/src/cli/commands/metadata.rs
@@ -1,3 +1,4 @@
+#![allow(clippy::unwrap_used)]
 use std::path::Path;
 
 use crate::cli::args::{MetadataArgs, MetadataSetArgs};
