+++ crates/utiles/src/cli/commands/rimraf.rs
@@ -1,3 +1,4 @@
+#![allow(clippy::unwrap_used)]
 use std::cell::Cell;
 
 use std::path::Path;
