+++ crates/utiles/src/cli/commands/touch.rs
@@ -1,14 +1,15 @@
 use std::path::Path;
 
+use chrono;
+use tracing::{debug, info};
+
 use crate::cli::args::TouchArgs;
 use crate::errors::UtilesResult;
 use crate::mbt::MbtType;
-use crate::sqlite::SqliteError;
-use crate::sqlite::{is_valid_page_size, Sqlike3};
-use crate::utilesqlite::Mbtiles;
+use crate::sqlite::is_valid_page_size;
+use crate::sqlite::{Sqlike3Async, SqliteError};
+use crate::utilesqlite::{MbtilesAsync, MbtilesAsyncSqliteClient};
 use crate::UtilesError;
-use chrono;
-use tracing::{debug, info};
 
 fn check_page_size(page_size: i64) -> UtilesResult<i64> {
     if is_valid_page_size(page_size) {
@@ -19,9 +20,8 @@ fn check_page_size(page_size: i64) -> UtilesResult<i64> {
     }
 }
 
-pub fn touch_main(args: &TouchArgs) -> UtilesResult<()> {
+pub async fn touch_main(args: &TouchArgs) -> UtilesResult<()> {
     let filepath = &args.filepath;
-
     let page_size = check_page_size(args.page_size.unwrap_or(4096))?;
     debug!("touch: {}", filepath);
     let dbtype: MbtType = args.mbtype();
@@ -32,20 +32,40 @@ pub fn touch_main(args: &TouchArgs) -> UtilesResult<()> {
     if fpth.extension().is_none() {
         Err(UtilesError::NoFspathExtension(filepath.to_string()))
     } else {
-        let filename_no_ext = fpth.file_stem().unwrap().to_str().unwrap();
-
-        assert!(!fpth.exists(), "Already exists: {}", fpth.display());
-        let mbtiles = Mbtiles::create(filepath, Some(dbtype))?;
-        info!("Created mbtiles: {:?}", filepath);
-        mbtiles.metadata_set("name", filename_no_ext)?;
-        mbtiles.metadata_set("mbtype", dbtype_str.to_string().as_str())?;
-        // current iso datetimestamp
-        let now = chrono::Utc::now();
-        let now_str = now.to_rfc3339();
-        mbtiles.metadata_set("ctime", now_str.as_str())?;
-        mbtiles.metadata_set("mtime", now_str.as_str())?;
-        mbtiles.pragma_page_size_set(page_size)?;
-        mbtiles.vacuum()?;
-        Ok(())
+        let stem_str = fpth.file_stem();
+        if stem_str.is_none() {
+            Err(UtilesError::NoFspathExtension(filepath.to_string()))
+        } else {
+            let filename_no_ext = stem_str;
+            match filename_no_ext {
+                Some(filename_no_ext) => {
+                    let filename_no_ext = {
+                        match filename_no_ext.to_str() {
+                            Some(filename_no_ext) => filename_no_ext.to_string(),
+                            None => {
+                                return Err(UtilesError::PathConversionError(
+                                    filepath.to_string(),
+                                ));
+                            }
+                        }
+                    };
+                    let mbtiles =
+                        MbtilesAsyncSqliteClient::open_new(filepath, Some(dbtype))
+                            .await?;
+                    info!("Created mbtiles: {:?}", filepath);
+                    mbtiles.metadata_set("name", &filename_no_ext).await?;
+                    mbtiles.metadata_set("mbtype", dbtype_str.as_str()).await?;
+                    // current iso datetimestamp
+                    let now = chrono::Utc::now();
+                    let now_str = now.to_rfc3339();
+                    mbtiles.metadata_set("ctime", now_str.as_str()).await?;
+                    mbtiles.metadata_set("mtime", now_str.as_str()).await?;
+                    mbtiles.pragma_page_size_set(page_size).await?;
+                    mbtiles.vacuum().await?;
+                    Ok(())
+                }
+                None => Err(UtilesError::NoFspathStem(filepath.to_string())),
+            }
+        }
     }
 }
